(#56, #39) Remove `MediaStore` and `Environment` APIs
