+++ android/src/main/kotlin/io/lakscastro/sharedstorage/common/ActivityListener.kt
@@ -1,4 +1,4 @@
-package io.lakscastro.sharedstorage.plugin
+package io.lakscastro.sharedstorage.common
 
 /**
  * Interface shared across API classes to make intuitive and clean [init] and [dispose] plugin
