+++ android/src/main/kotlin/io/lakscastro/sharedstorage/common/Listenable.kt
@@ -1,4 +1,4 @@
-package io.lakscastro.sharedstorage.plugin
+package io.lakscastro.sharedstorage.common
 
 import io.flutter.plugin.common.BinaryMessenger
