+++ android/src/main/kotlin/io/lakscastro/sharedstorage/plugin/SharedStorageApi.kt
@@ -1,6 +1,8 @@
-package io.lakscastro.sharedstorage.storageaccessframework
+package io.lakscastro.sharedstorage.plugin
 
+import android.content.ActivityNotFoundException
 import android.content.Intent
+import android.graphics.Point
 import android.net.Uri
 import android.os.Build
 import android.provider.DocumentsContract
@@ -9,11 +11,10 @@ import androidx.annotation.RequiresApi
 import androidx.documentfile.provider.DocumentFile
 import com.anggrayudi.storage.file.child
 import io.flutter.plugin.common.*
-import io.flutter.plugin.common.EventChannel.StreamHandler
 import io.lakscastro.sharedstorage.ROOT_CHANNEL
 import io.lakscastro.sharedstorage.SharedStoragePlugin
-import io.lakscastro.sharedstorage.plugin.*
-import io.lakscastro.sharedstorage.storageaccessframework.lib.*
+import io.lakscastro.sharedstorage.common.*
+import io.lakscastro.sharedstorage.plugin.lib.*
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -23,20 +24,16 @@ import java.io.InputStream
 import java.io.OutputStream
 
 /**
- * Aimed to implement strictly only the APIs already available from the native and original
- * `DocumentFile` API
- *
- * Basically, this is just an adapter of the native `DocumentFile` class to a Flutter Plugin class,
- * without any modifications or abstractions
+ * APIs implementations are contained in this class and used by [SharedStoragePlugin]
  */
-internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
-    MethodChannel.MethodCallHandler,
-    PluginRegistry.ActivityResultListener,
-    Listenable,
-    ActivityListener,
-    StreamHandler {
+class SharedStorageApi(val plugin: SharedStoragePlugin) :
+  MethodChannel.MethodCallHandler,
+  PluginRegistry.ActivityResultListener,
+  Listenable,
+  ActivityListener,
+  EventChannel.StreamHandler {
   private val pendingResults: MutableMap<Int, Pair<MethodCall, MethodChannel.Result>> =
-      mutableMapOf()
+    mutableMapOf()
   private var channel: MethodChannel? = null
   private var eventChannel: EventChannel? = null
   private var eventSink: EventChannel.EventSink? = null
@@ -47,6 +44,41 @@ internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
 
   override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
     when (call.method) {
+      GET_DOCUMENT_THUMBNAIL -> {
+        if (Build.VERSION.SDK_INT >= API_21) {
+          val uri = Uri.parse(call.argument("uri"))
+          val width = call.argument<Int>("width")!!
+          val height = call.argument<Int>("height")!!
+
+          val bitmap =
+            DocumentsContract.getDocumentThumbnail(
+              plugin.context.contentResolver,
+              uri,
+              Point(width, height),
+              null
+            )
+
+          CoroutineScope(Dispatchers.Default).launch {
+            if (bitmap != null) {
+              val base64 = bitmapToBase64(bitmap)
+
+              val data =
+                mapOf(
+                  "base64" to base64,
+                  "uri" to "$uri",
+                  "width" to bitmap.width,
+                  "height" to bitmap.height,
+                  "byteCount" to bitmap.byteCount,
+                  "density" to bitmap.density
+                )
+
+              launch(Dispatchers.Main) { result.success(data) }
+            }
+          }
+        } else {
+          result.notSupported(call.method, API_21)
+        }
+      }
       GET_DOCUMENT_CONTENT -> {
         val uri = Uri.parse(call.argument<String>("uri")!!)
 
@@ -61,9 +93,9 @@ internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
         }
       }
       OPEN_DOCUMENT_TREE ->
-          if (Build.VERSION.SDK_INT >= API_21) {
-            openDocumentTree(call, result)
-          }
+        if (Build.VERSION.SDK_INT >= API_21) {
+          openDocumentTree(call, result)
+        }
       CREATE_FILE ->
           if (Build.VERSION.SDK_INT >= API_21) {
             createFile(
@@ -115,48 +147,48 @@ internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
           result.success(documentFromUri(plugin.context, uri)?.canRead())
         }
       LENGTH ->
-          if (Build.VERSION.SDK_INT >= API_21) {
-            result.success(
-              documentFromUri(
-                plugin.context,
-                call.argument<String?>("uri") as String
-              )?.length()
-            )
-          }
-      EXISTS ->
-          if (Build.VERSION.SDK_INT >= API_21) {
-            result.success(
-              documentFromUri(
-                plugin.context,
-                call.argument<String?>("uri") as String
-              )?.exists()
-            )
-          }
-      DELETE ->
-          if (Build.VERSION.SDK_INT >= API_21) {
-            result.success(
-              documentFromUri(
-                plugin.context,
-                call.argument<String?>("uri") as String
-              )?.delete()
-            )
-          }
-      LAST_MODIFIED ->
-          if (Build.VERSION.SDK_INT >= API_21) {
-            val document = documentFromUri(
+        if (Build.VERSION.SDK_INT >= API_21) {
+          result.success(
+            documentFromUri(
               plugin.context,
               call.argument<String?>("uri") as String
-            )
+            )?.length()
+          )
+        }
+      EXISTS ->
+        if (Build.VERSION.SDK_INT >= API_21) {
+          result.success(
+            documentFromUri(
+              plugin.context,
+              call.argument<String?>("uri") as String
+            )?.exists()
+          )
+        }
+      DELETE ->
+        if (Build.VERSION.SDK_INT >= API_21) {
+          result.success(
+            documentFromUri(
+              plugin.context,
+              call.argument<String?>("uri") as String
+            )?.delete()
+          )
+        }
+      LAST_MODIFIED ->
+        if (Build.VERSION.SDK_INT >= API_21) {
+          val document = documentFromUri(
+            plugin.context,
+            call.argument<String?>("uri") as String
+          )
 
-            result.success(document?.lastModified())
-          }
+          result.success(document?.lastModified())
+        }
       CREATE_DIRECTORY -> {
         if (Build.VERSION.SDK_INT >= API_21) {
           val uri = call.argument<String?>("uri") as String
           val displayName = call.argument<String?>("displayName") as String
 
           val createdDirectory =
-              documentFromUri(plugin.context, uri)?.createDirectory(displayName) ?: return
+            documentFromUri(plugin.context, uri)?.createDirectory(displayName) ?: return
 
           result.success(createDocumentFileMap(createdDirectory))
         } else {
@@ -239,14 +271,56 @@ internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
 
         if (Build.VERSION.SDK_INT >= API_21) {
           val document = documentFromUri(plugin.context, uri)
-          val childDocument = document?.child(plugin.context, path, requiresWriteAccess)
+          val childDocument =
+            document?.child(plugin.context, path, requiresWriteAccess)
 
           result.success(createDocumentFileMap(childDocument))
         } else {
           result.notSupported(CHILD, API_21, mapOf("uri" to uri))
         }
       }
+      OPEN_DOCUMENT_FILE -> openDocumentFile(call, result)
       else -> result.notImplemented()
+    }
+  }
+
+  private fun openDocumentFile(call: MethodCall, result: MethodChannel.Result) {
+    val uri = Uri.parse(call.argument<String>("uri")!!)
+    val type =
+      call.argument<String>("type") ?: plugin.context.contentResolver.getType(
+        uri
+      )
+
+    val intent =
+      Intent(Intent.ACTION_VIEW).apply {
+        flags = Intent.FLAG_GRANT_READ_URI_PERMISSION
+        data = uri
+      }
+
+    try {
+      plugin.binding?.activity?.startActivity(intent, null)
+
+      Log.d("sharedstorage", "Successfully launched uri $uri ")
+
+      result.success(null)
+    } catch (e: ActivityNotFoundException) {
+      result.error(
+        EXCEPTION_ACTIVITY_NOT_FOUND,
+        "There's no activity handler that can process the uri $uri of type $type",
+        mapOf("uri" to "$uri", "type" to type)
+      )
+    } catch (e: SecurityException) {
+      result.error(
+        EXCEPTION_CANT_OPEN_FILE_DUE_SECURITY_POLICY,
+        "Missing read and write permissions for uri $uri of type $type to launch ACTION_VIEW activity",
+        mapOf("uri" to "$uri", "type" to "$type")
+      )
+    } catch (e: Throwable) {
+      result.error(
+        EXCEPTION_CANT_OPEN_DOCUMENT_FILE,
+        "Couldn't start activity to open document file for uri: $uri",
+        mapOf("uri" to "$uri")
+      )
     }
   }
 
@@ -257,24 +331,24 @@ internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
     val initialUri = call.argument<String>("initialUri")
 
     val intent =
-        Intent(Intent.ACTION_OPEN_DOCUMENT_TREE).apply {
-          addFlags(
-              if (grantWritePermission) Intent.FLAG_GRANT_WRITE_URI_PERMISSION
-              else Intent.FLAG_GRANT_READ_URI_PERMISSION
-          )
+      Intent(Intent.ACTION_OPEN_DOCUMENT_TREE).apply {
+        addFlags(
+          if (grantWritePermission) Intent.FLAG_GRANT_WRITE_URI_PERMISSION
+          else Intent.FLAG_GRANT_READ_URI_PERMISSION
+        )
 
-          if (initialUri != null) {
-            val tree = DocumentFile.fromTreeUri(plugin.context, Uri.parse(initialUri))
+        if (initialUri != null) {
+          val tree = DocumentFile.fromTreeUri(plugin.context, Uri.parse(initialUri))
 
-            if (Build.VERSION.SDK_INT >= API_26) {
-              putExtra(
-                if (Build.VERSION.SDK_INT >= API_26) DocumentsContract.EXTRA_INITIAL_URI
-                else DOCUMENTS_CONTRACT_EXTRA_INITIAL_URI,
-                tree?.uri
-              )
-            }
+          if (Build.VERSION.SDK_INT >= API_26) {
+            putExtra(
+              if (Build.VERSION.SDK_INT >= API_26) DocumentsContract.EXTRA_INITIAL_URI
+              else DOCUMENTS_CONTRACT_EXTRA_INITIAL_URI,
+              tree?.uri
+            )
           }
         }
+      }
 
     if (pendingResults[OPEN_DOCUMENT_TREE_CODE] != null) return
 
@@ -285,11 +359,11 @@ internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
 
   @RequiresApi(API_21)
   private fun createFile(
-      result: MethodChannel.Result,
-      mimeType: String,
-      displayName: String,
-      directory: String,
-      content: ByteArray
+    result: MethodChannel.Result,
+    mimeType: String,
+    displayName: String,
+    directory: String,
+    content: ByteArray
   ) {
     createFile(Uri.parse(directory), mimeType, displayName, content) {
       result.success(createDocumentFileMap(this))
@@ -298,11 +372,11 @@ internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
 
   @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
   private fun createFile(
-      treeUri: Uri,
-      mimeType: String,
-      displayName: String,
-      content: ByteArray,
-      block: DocumentFile?.() -> Unit
+    treeUri: Uri,
+    mimeType: String,
+    displayName: String,
+    content: ByteArray,
+    block: DocumentFile?.() -> Unit
   ) {
     val createdFile = documentFromUri(plugin.context, treeUri)!!.createFile(mimeType, displayName)
 
@@ -342,24 +416,24 @@ internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
     val persistedUriPermissions = plugin.context.contentResolver.persistedUriPermissions
 
     result.success(
-        persistedUriPermissions
-            .map {
-              mapOf(
-                  "isReadPermission" to it.isReadPermission,
-                  "isWritePermission" to it.isWritePermission,
-                  "persistedTime" to it.persistedTime,
-                  "uri" to "${it.uri}"
-              )
-            }
-            .toList()
+      persistedUriPermissions
+        .map {
+          mapOf(
+            "isReadPermission" to it.isReadPermission,
+            "isWritePermission" to it.isWritePermission,
+            "persistedTime" to it.persistedTime,
+            "uri" to "${it.uri}"
+          )
+        }
+        .toList()
     )
   }
 
   @RequiresApi(API_19)
   private fun releasePersistableUriPermission(result: MethodChannel.Result, directoryUri: String) {
     plugin.context.contentResolver.releasePersistableUriPermission(
-        Uri.parse(directoryUri),
-        Intent.FLAG_GRANT_WRITE_URI_PERMISSION
+      Uri.parse(directoryUri),
+      Intent.FLAG_GRANT_WRITE_URI_PERMISSION
     )
 
     result.success(null)
@@ -378,9 +452,9 @@ internal class DocumentFileApi(private val plugin: SharedStoragePlugin) :
 
           if (uri != null) {
             plugin.context.contentResolver.takePersistableUriPermission(
-                uri,
-                if (grantWritePermission) Intent.FLAG_GRANT_WRITE_URI_PERMISSION
-                else Intent.FLAG_GRANT_READ_URI_PERMISSION
+              uri,
+              if (grantWritePermission) Intent.FLAG_GRANT_WRITE_URI_PERMISSION
+              else Intent.FLAG_GRANT_READ_URI_PERMISSION
             )
 
             pendingResult.second.success("$uri")
