+++ example/lib/screens/file_explorer/file_explorer_card.dart
@@ -5,7 +5,7 @@ import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:shared_storage/saf.dart';
+import 'package:shared_storage/shared_storage.dart';
 
 import '../../theme/spacing.dart';
 import '../../widgets/buttons.dart';
