+++ example/lib/screens/granted_uris/granted_uri_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:shared_storage/saf.dart';
+import 'package:shared_storage/shared_storage.dart';
 
 import '../../theme/spacing.dart';
 import '../../widgets/buttons.dart';
