--- lib/environment.dart
@@ -1,4 +1,0 @@
-library shared_storage;
-
-export './src/environment/environment.dart';
-export './src/environment/environment_directory.dart';
