--- lib/media_store.dart
@@ -1,4 +1,0 @@
-library shared_storage;
-
-export './src/media_store/media_store.dart';
-export './src/media_store/media_store_collection.dart';
