--- lib/saf.dart
@@ -1,8 +1,0 @@
-library shared_storage;
-
-export './src/saf/document_bitmap.dart';
-export './src/saf/document_file.dart';
-export './src/saf/document_file_column.dart';
-export './src/saf/partial_document_file.dart';
-export './src/saf/saf.dart';
-export './src/saf/uri_permission.dart';
