+++ lib/shared_storage.dart
@@ -1,3 +1,8 @@
-export './environment.dart';
-export './media_store.dart';
-export './saf.dart';
+library shared_storage;
+
+export './src/api/document_bitmap.dart';
+export './src/api/document_file.dart';
+export './src/api/document_file_column.dart';
+export './src/api/partial_document_file.dart';
+export './src/api/saf.dart';
+export './src/api/uri_permission.dart';
