+++ lib/src/api/document_bitmap.dart
