+++ lib/src/common/annotation.dart
@@ -1,0 +1,4 @@
+const deprecated = Deprecated(
+  'This method should no longer be used, and will be removed in a near release'
+  '\nRefer to the documentation https://lakscastro.github.io/shared-storage to see how to migrate',
+);
