+++ lib/src/common/invoke_map_method.dart
@@ -1,5 +1,5 @@
+import '../api/document_file.dart';
 import '../channels.dart';
-import 'document_file.dart';
 
 /// Helper method to invoke a native SAF method and return a document file
 /// if not null, shouldn't be called directly from non-package code
