+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: shared_storage
 description: "Flutter plugin to work with external storage."
-version: 0.4.1
+version: 0.5.0
 homepage: https://github.com/lakscastro/shared-storage
 repository: https://github.com/lakscastro/shared-storage
 issue_tracker: https://github.com/lakscastro/shared-storage/issues
